remove parse schema code and organize code
